Make a list of grouping tags when filtering to delete AMI.
The previous code make an wrong pression that it is safe as we always keep 1
last ami. However this is only correct for the current filters tag.

And if you put filter tag like the example in the code which is too generic
then it would remove ami unexpectedly.

Luckily that for ami with current running ec2 it never touch. However for these
amis which we use business_hour rules which does not run at non business hours,
the deregister logic will remove them as well.

This change will make a list of filter group and calling the deregister
function one by one. The example shoes more level of filtering to be sure the
hint of usage is clear - you have to pass more details into the filtering to
get the correct outcome.